+++ examples/delete-all-logs.example.ts
@@ -1,0 +1,4 @@
+import { logger } from "./index.example.js";
+
+// If we provide not options it will delete all file logs content
+await logger.deleteLogs();