+++ examples/read-logs-by-severity.example.ts
@@ -1,5 +1,5 @@
 import { LogSeverity } from "../src/index.js";
-import { logger } from "./index.example";
+import { logger } from "./index.example.js";
 
 const logs = await logger.getLogsBySeverity(LogSeverity.debug);
 console.log(logs);